requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.4.0
django-js-asset==2.2.0
django-tinymce==4.1.0
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.2
